feat(rank): add gamemode ranking system and fix map rating calculation
Add FloodWonEvent and ranking classes for flood, attack, and survival gamemodes to track and display best clear times per map. The system stores records in persistent settings and announces new best times. Also fix map rating calculation by changing totalScore to float to prevent integer division issues.

+++ plugin/src/main/java/plugin/service/MapRating.java
@@ -24,7 +24,7 @@ public class MapRating {
 
         public MapRatingStat stat() {
             int totalVotes = 0;
-            int totalScore = 0;
+            float totalScore = 0;
 
             for (int i = 0; i < 5; i++) {
                 int count = stars[i];
